Cria o grupo de recursos e corrige a conexão bd
